requirements.txt
pdfplumber==0.10.3
PyPDF2==3.0.1
reportlab==4.0.7
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0
faiss-cpu==1.7.4
numpy==1.24.3
xgboost==2.0.3
scikit-learn==1.3.2
fastapi==0.108.0
uvicorn==0.25.0
streamlit==1.29.0
python-multipart==0.0.6
pandas==2.1.4
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0
tqdm==4.66.1
